# Upgrade to serde 1.0 by implementing serialization directly