--- a/Cargo.toml
+++ b/Cargo.toml
@@ -12,12 +12,12 @@ license = "MIT"
 build = "build.rs"
 
 [dependencies]
-serde = "^0.8"
-serde_json = "^0.8"
+serde = { version = "1", features = ["rc"] }
+serde_json = "1"
 lazy_static = "^0.2.2"
 
 [build-dependencies]
-serde_json = "^0.8"
+serde_json = "1"
 slug = "0.1.2"
 
 [features]