--- a/jmespath-cli/Cargo.toml
+++ b/jmespath-cli/Cargo.toml
@@ -8,8 +8,8 @@ license = "MIT"
 name = "jp"
 
 [dependencies]
-serde = "^0.8"
-serde_json = "^0.8"
+serde = "1"
+serde_json = "1"
 clap = "^2.4"
 
 [dependencies.jmespath]